* config/linux/x86/futex_bits.h (sys_futex0) [__x86_64__]: Change 	operands "op" and "val" to int.
	* config/linux/sh/futex_bits.h (sys_futex0) Change operands
	"op" and "val" to int.

	* config/linux/alpha/futex_bits.h (sys_futex0) Change operands
	"op" and "val" to int.

	* config/linux/futex.cc (gtm_futex_wait, gtm_futex_wake):
	Declare as static int.
	(FUTEX_PRIVATE_FLAG): Remove L suffix.
	* config/linux/futex_bits.h (sys_futex0) Change operand "op" to int.

	Revert:
	* config/linux/x86/futex.h (sys_futex0) [!__x86_64__]:
	Change operand "op" to long.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@223844 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libitm/config/linux/x86/futex_bits.h
@@ -28,7 +28,7 @@
 # endif
 
 static inline long
-sys_futex0 (std::atomic<int> *addr, long op, long val)
+sys_futex0 (std::atomic<int> *addr, int op, int val)
 {
   register long r10 __asm__("%r10") = 0;
   long res;
@@ -47,7 +47,7 @@ sys_futex0 (std::atomic<int> *addr, long op, long val)
 # endif
 
 static inline long
-sys_futex0 (std::atomic<int> *addr, long op, int val)
+sys_futex0 (std::atomic<int> *addr, int op, int val)
 {
   long res;
